Create readme
Please run it on local wamp server.
Index.php is a default page where you can write the string that will be base64 encoded and will sent to calculator where result will be calculated.

Calculus.php shows result of the expression or history of the five latest expressions according to the end point called. 

On Index.php a history link has been provided which will show a JSON formatted history. If expression will be empty the error will be shown in JSON format otherwise the result of the expression will be shown.